Add files via upload
Hopefully the final MuGet update. Create a class that derives from MuGet to create your widget (look into MuGetSlider to get some inspiration). To test it just make a windows form and put the lines:
            MuGet muGet = new MuGetSlider("bla", 25, 25, 150, 150);
            this.Controls.Add(muGet);
in the constructor to place your widget. Change the 25s to change the position and the 150s to change the size.