Add files via upload
SoundEngine now usable to create Loops.
The Application needs one SoundEngine with wich all widgets may interact. Widgets need to store a reference.

The constructor is a mess (I am trying out stuff there.). But the loop methods should work.
With startPlayback the microfone input will be mapped to the speakers which can be stopped with stopPlayback.